Added Unity Netcode for GameObjects and Lobby Services, replaced Photon Bolt

+++ Assets/Scripts/Core/Spells/Spell Processing/SpellDestinationEntry.cs
@@ -5,7 +5,7 @@ namespace Core
     internal class SpellDestinationEntry
     {
         public Vector3 Destination { get; set; }
-        public int Delay { get; set; }
+        public float Delay { get; set; }
         public bool Processed { get; set; }
         public int EffectMask { get; set; }
     }
